+++ precompiles/vesting/utils_test.go
@@ -30,7 +30,7 @@ import (
 	"github.com/evmos/evmos/v15/utils"
 	"github.com/evmos/evmos/v15/x/evm/statedb"
 	evmtypes "github.com/evmos/evmos/v15/x/evm/types"
-	inflationtypes "github.com/evmos/evmos/v15/x/inflation/types"
+	inflationtypes "github.com/evmos/evmos/v15/x/inflation/v1/types"
 	vestingtypes "github.com/evmos/evmos/v15/x/vesting/types"
 
 	//nolint:revive // dot imports are fine for Ginkgo