+++ x/erc20/keeper/ibc_callbacks_test.go
@@ -27,7 +27,7 @@ import (
 	"github.com/evmos/evmos/v15/contracts"
 	claimstypes "github.com/evmos/evmos/v15/x/claims/types"
 	"github.com/evmos/evmos/v15/x/erc20/types"
-	inflationtypes "github.com/evmos/evmos/v15/x/inflation/types"
+	inflationtypes "github.com/evmos/evmos/v15/x/inflation/v1/types"
 	vestingtypes "github.com/evmos/evmos/v15/x/vesting/types"
 )